domain분리
