+++ src/main/java/racing/domain/InitCars.java
@@ -1,0 +1,6 @@
+package racing.domain;
+
+public class InitCars {
+
+
+}
