+++ src/main/java/racing/domain/MoveCars.java
@@ -1,0 +1,21 @@
+package racing.domain;
+
+import racing.Car;
+
+import static racing.TheGame.cars;
+import static racing.TheGame.random;
+
+public class MoveCars {
+
+    static int moveCars() {
+        int carPosistion = 0;
+
+        for (Car car : cars) {
+            /* carPosistion = Math.max(car.move(random.nextInt(10)), carPosistion);*/
+            carPosistion = car.move(random.nextInt(10));
+        }
+
+        return carPosistion;
+    }
+
+}
